+++ demo/wx/app.js
@@ -1,8 +1,5 @@
 // app.js
 App({
-  onLaunch() {
-    // wx.cloud.init();
-  },
   onPageNotFound() {
     wx.redirectTo({
       url: "/pages/index/index",