+++ demo/wx/components/parser/libs/config.js
@@ -61,16 +61,13 @@ module.exports = {
   trustTags: makeMap('a,abbr,ad,audio,b,blockquote,br,code,col,colgroup,dd,del,dl,dt,div,em,fieldset,h1,h2,h3,h4,h5,h6,hr,i,img,ins,label,legend,li,ol,p,q,source,span,strong,sub,sup,table,tbody,td,tfoot,th,thead,tr,title,ul,video' + (canIUse ? ',bdi,bdo,caption,pre,rt,ruby' : '')),
   // 默认的标签样式
   userAgentStyles: {
-    a: 'color:#366092;word-break:break-all;padding:1.5px 0 1.5px 0',
     address: 'font-style:italic',
     big: 'display:inline;font-size:1.2em',
     blockquote: 'background-color:#f6f6f6;border-left:3px solid #dbdbdb;color:#6c6c6c;padding:5px 0 5px 10px',
     center: 'text-align:center',
     cite: 'font-style:italic',
     dd: 'margin-left:40px',
-    img: 'max-width:100%',
     mark: 'background-color:yellow',
-    picture: 'max-width:100%',
     pre: 'font-family:monospace;white-space:pre;overflow:scroll',
     s: 'text-decoration:line-through',
     small: 'display:inline;font-size:0.8em',