+++ demo/wx/pages/demo/demo.js
@@ -9,7 +9,7 @@ Page({
     this.anchor = e.anchor;
   },
   // 富文本渲染完成
-  ready() {
+  ready(e) {
     if (this.anchor)
       this.context.navigateTo({
         id: this.anchor
@@ -43,80 +43,6 @@ Page({
         })
       }
     }
-  },
-  // 图片长按事件
-  imglongtap(e) {
-    this.url = e.detail.src;
-    this.setData({
-      showActionsheet: true,
-      groups: [{
-        text: '保存到手机',
-        value: 0
-      }, {
-        text: '复制图片地址',
-        value: 1
-      }]
-    })
-    // 云调用 img.scanQRCode
-    wx.cloud.callFunction({
-      name: 'scanImg',
-      data: {
-        src: this.url
-      },
-      success: (res) => {
-        if (res.result.codeResults.length) {
-          this.scanRes = res.result.codeResults[0].data;
-          this.setData({
-            groups: [{
-              text: '保存到手机',
-              value: 0
-            }, {
-              text: '复制图片地址',
-              value: 1
-            }, {
-              text: '复制识别结果',
-              value: 2
-            }]
-          })
-        }
-      }
-    })
-  },
-  // 图片菜单点击
-  actiontap(e) {
-    if (e.detail.value == 0) {
-      function saveToAlbum() {
-        wx.downloadFile({
-          url: this.url,
-          success(res) {
-            wx.saveImageToPhotosAlbum({
-              filePath: res.tempFilePath
-            })
-          }
-        })
-      }
-      wx.getSetting({
-        success(res) {
-          if (!res.authSetting['scope.writePhotosAlbum']) {
-            wx.authorize({
-              scope: 'scope.writePhotosAlbum',
-              success: saveToAlbum
-            })
-          } else
-            saveToAlbum();
-        }
-      })
-    } else if (e.detail.value == 1)
-      wx.setClipboardData({
-        data: this.url
-      })
-    else
-      wx.setClipboardData({
-        data: this.scanRes
-      })
-    this.setData({
-      showActionsheet: false
-    })
   },
   // 页面分享
   onShareAppMessage() {