+++ docs/changelog.md
@@ -1,13 +1,27 @@
 ## 更新日志 {docsify-ignore} ##
 
+#### 2020.03.17 beta ####
+1. `U` 通过 `image`（经过一些处理后）来显示图片（替代 `rich-text`），可以实现以下优化（仅微信包）：  
+   1. `2.3.0` 起支持云文件 `ID`  
+   2. `2.7.0` 起支持长按弹出菜单（可以识别小程序码，同时去除了 `imglongtap` 事件）  
+   3. `2.9.0` 起支持 `webp` 图片  
+   4. 使用 `image` 原生的 `lazy-load`，可能具有更好的性能  
+   5. 加载错误时能够触发 `error` 事件，且可以重设 `src` [详细](/instructions#事件)  
+   
+   可能存在的问题：  
+   若没有设置大小图片会在加载完成后突然从默认大小（300 × 100）变为原大小（图片较大，加载较慢时较明显），可以在 `trees.wxss` 的 `._img` 中调整默认大小  
+2. `U` `a` 标签支持 `:visited` 效果（默认变为紫色，可在 `trees.wxss` 中调整）  
+3. `F` 修复了 `a` 标签所在段落使用一些特殊实体编码时可能导致错误换行的问题 [详细](https://github.com/jin-yufeng/Parser/issues/87)  
+4. `F` 修复了 `uni-app` 包 `H5` 端在创建时设置数据无法显示的问题 [详细](https://github.com/jin-yufeng/Parser/issues/89)  
+  
+
 #### 2020.03.12 ####
 1. `A` 增加了 `compress` 属性，可以设置压缩等级 [详细](/instructions#compress)  
 2. `A` 配置项中增加了 `filter` 和 `onText` 方法，可以在解析过程中进行自定义处理 [详细](/instructions#配置项)  
 3. `A` 增加了 `rect` 的 `api`，可以获取内容的大小和位置 [详细](/instructions#rect)  
-4. `U` `picture` 标签中若设置 `webp` 的 `source`，将只有 `android` 端采用，可用于兼容 [详细](/features#多媒体多资源加载)  
-5. `U` `setContent` 的 `api` 支持传入 `append` 参数表示是否在尾部追加（用于加载更多）[详细](/instructions#setContent)  
-6. `U` 支持通过 `base` 标签设置主域名（同 `domain` 属性，但优先级更低）  
-7. `F` 修复了在 `ready` 事件触发前再次设置数据会导致 `ready` 事件不停触发的问题  
+4. `U` `setContent` 的 `api` 支持传入 `append` 参数表示是否在尾部追加（用于加载更多）[详细](/instructions#setContent)  
+5. `U` 支持通过 `base` 标签设置主域名（同 `domain` 属性，但优先级更低）  
+6. `F` 修复了在 `ready` 事件触发前再次设置数据会导致 `ready` 事件不停触发的问题  
 
 #### 2020.03.07 ####
 1. `A` 增加了 `preLoad` 的 `api`，可以预加载富文本中的图片 [详细](/instructions#preLoad)
@@ -34,7 +48,7 @@
 5. `U` `uni-app` 包 `APP(v3)` 端支持 `iframe` 标签  
 
 #### 2020.02.17 ####
-1. `A` 增加了 `imglongtap` 事件，图片被长按时触发，可用于显示自定义菜单 [详细](/instructions#事件)  
+1. `A` 增加了 `imglongtap` 事件，图片被长按时触发，可用于显示自定义菜单 *附：20200317版本中被删除，使用原生的长按菜单*  
 2. `U` 优化了双击缩放的效果
 3. `U` 图片设置的宽度超出屏幕宽度时自动将高度设置为 `auto`，避免变形（同时移除了 `img-mode` 属性）
 4. `U` 修改了部分文件和文件夹的命名（**引入路径有变化**）[详细](/instructions#在原生框架中使用)